refrefactor: refactor(files): extract reusable helper for converting pattern strings to globs
